Fix footer, fix logo not showing on certain pages
Removed sticky footer as footer was covering text on longer reports. To push footer to the bottom, added flexbox to body of page, pushed footer outside of body.

Added a '/' to the src of the logo, to use a root-relative path. The logo was not previously showing on pages that were more than 1 clicks deep